Fold observation registry into observations module.
Move create_observation_spec into observations.py so the PGC specs package is just actions and observations.
