+++ src/environments/PlantGrowthChamber/specs/observations.py
@@ -10,6 +10,7 @@ import joblib
 import numpy as np
 import pandas as pd
 
+from environments.PlantGrowthChamber.specs.actions import ActionSpec
 from environments.PlantGrowthChamber.utils import (
     get_one_hot_time_observation,
     hours_normalized,
@@ -313,3 +314,23 @@ class WallStatsEmbeddingObservation(ObservationSpec):
     def update_action_trace(self, action: Any, backend: Any) -> None:
         assert self.action_uema is not None
         self.action_uema.update(jnp.array(action))
+
+
+def create_observation_spec(
+    name: str, action_spec: ActionSpec, env_params: dict[str, Any]
+) -> ObservationSpec:
+    if name in ("scalar", "area"):
+        return AreaObservation()
+    if name == "timestamp":
+        return TimestampObservation()
+    if name == "one_hot_time":
+        return OneHotTimeObservation()
+    if name == "time_area":
+        return TimeAreaObservation()
+    if name == "time_dli":
+        return TimeDLIObservation()
+    if name == "day_area_trace":
+        return DayAreaTraceObservation(trace_dim=action_spec.trace_dim)
+    if name == "wall_stats_embedding":
+        return WallStatsEmbeddingObservation(trace_dim=action_spec.trace_dim)
+    raise ValueError(f"Unknown observation spec: {name}")
